Implement Freivalds' algorithm for matrix products
Freivalds' algorithm verifies that a matrix C is the product of two
matrices A and B in quadratic time by comparing the evaluation of C as a
linear map with the consecutive evaluations of B and then A.  Here,
evaluation is on a vector consisting of powers of a randomly chosen
field element.  This choice of evaluation vector allows the analysis to
invoke the Schwartz-Zippel lemma to prove that the algorithm correctly
identifies an incorrect matrix product with high probability when
working over a large finite field.

The binary freivald.rs implements Freivalds' algorithm and tests it with
randomly chosen matrices A and B for a few correct and incorrect values
of the matrix C.