Implement threaded IP framework and test protocol
Implements a simple framework for specifying and running interactive
protocols (IPs) via Rust threads.  Message passing is handled by a
custom bidirectional logged channel type using an asynchronous send
command and a blocking receive command.  Data which is sent down the
channel is cloned in order to simplify implementations using the
framework.

An example IP is implemented in add_one.rs.  In this protocol, the
verifier sends a number to the prover to determine whether the prover is
capable of adding one to it.  If the prover successfully returns the
sent number plus one, the verifier accepts, and otherwise the verifier
rejects.  This binary first tests a correct prover which adds one, and
then tests an incorrect prover which responds with a random number.