Implement several "small" fields useful for examples
This commit implements several "small" finite fields using the Arkworks
PrimeField trait and the Fp64 concrete implementation.  These fields are
useful for demonstrating the functionality of algorithms and protocols
in an easy-to-read way, as well as for showcasing situations where
success probability depends concretely on the cardinality of an
underlying field.  Three fields F5, F13, and F251 are implemented, and a
binary is included which exercises these fields.

An important implementation detail helpful for understanding these field
types is that the field elements for Arkworks Fp objects are represented
using the so-called "montgomery form", which premultiplies elements by a
fixed constant R in order to make modular multiplication more efficient.
We highlight the parameters represented in this way with an inline
comment.