Revise interactive protocol framework
The interactive protocol framework was revised to decouple the messaging
and logging parts of the process.  Messaging is implemented simply as a
bidirectional channel between IP participants (without any built-in
logging), and logging is implemented as a separate Log object which is
used explicitly to record Strings in a shared Vec which can be
referenced later.  This adds some extra work for concrete
implementations, but also allows more flexible "annotations" describing
what a protocol is doing beyond just the record of passed messages.

The "add_one" binary is updated to make use of the new implementation,
and demonstrates how you might provide log entries to describe the
operation of an IP participant more verbosely.