Revise Freivalds' algorithm binary
Made a few changes to the binary demonstrating Freivalds' algorithm.
First, the underlying finite field was changed to F5 to better
demonstrate that the algorithm may return an incorrect result with
probability depending on the size of the field.  Second, the Decision
enum was changed to a simpler tuple struct.