Implement polynomial utility functions
Added a module with functions to help with a few high-level operations
for working with Arkware polynomials.  A core function which deserves
some comment is `polynomials::partial_summation`, which is a multitool
for computing partial evaluations and summations for a multivariate
polynomial represented as a sparse list of monomials.

The function takes as input a multivariate polynomial `p` and a list
whose entries each correspond with a variable of `p`.  A value of `None`
in this list indicates that the variable should be left unevaluated,
while a value of `Some(Vec<F>)` indicates that the variable should be
summed over the field elements in the Vec.  Note that if only a single
field element is provided in the latter case, this is equivalent to
partial evaluation of the variable at that field element.  Applying all
of the specified summations requires only a single iteration over the
list of monomials.