Implement PAZK Protocol 11 for polynomial commitment
Protocol 11 of PAZK gives a sigma protocol allowing a prover to convince a
verifier that they know openings of two Pedersen commitments representing,
respectively, a vector of finite field elements and its inner product with a
specified public vector.  Using the vector of powers of a fixed field
elements, this gives a univariate polynomial commitment scheme, albeit one
which is not succinct since the size of the opening information and the
verifier complexity is linear in the length of the vectors involved.

The binary `pedersen_poly_commitment` executes Protocol 11 for the evaluation
of a random polynomial of specified degree at a random evaluation point.  A
detailed transcript of protocol execution is printed to standard output; the
group for Pedersen commitments, and its associated scalar field, are chosen
to be small so that the output is easy to read and the computations can be
checked by hand.