Small updates to group_utils library file
Changed the name of the function for multi-exponentiation in cryptographic
groups to `msm` to match naming convention in Arkworks, and generalized its
signature to allow for any classes defining the required arithmetic
properties.  Also, changed the signature of the `list_vec` function to
specify a separator string.  Updated all code which made use of these
functionalities to reflect the modified signatures.