Implement Bulletproof inner product protocol
The Bulletproof protocol allows a prover to convince a verifier that the inner
product of a vector represented as a generalized Pedersen commitment (or other
homomorphic commitment) has a given inner product with a specified public
vector, using only logarithmic communication, but with linear verifier time.

The binary file `bulletproof.rs` implements the Bulletproof protocol and
produces a transcript of the protocol being employed to convince a verifier
that a polynomial with coefficients encoded in a generalized Pedersen
commitment evaluates to some value at a specified evaluation point.  The
protocol uses a small elliptic curve group with small associated scalar field
to produce a transcript which can be reasoned about by hand.